refactor: processTranslationCall
